finished login and signup logic
