+++ views/homepage.handlebars
@@ -1,1 +1,18 @@
 <h1>I am the home page</h1>
+
+{{#each horses}}
+<div class="card">
+    <div class="card-content">
+        <p class="title">
+            <a href="/horseCharacter/{{id}}"></a>
+        </p>
+        <p class="subtitle">By: {{horseCharacter.horse_name}}</p>
+        <p class="subtitle">{{horseCharacter.horse_breed}}</p>
+        <p class="subtitle">{{horseCharacter.horse_power}}</p>
+        <p class="subtitle">{{horseCharacter.horse_speed}}</p>
+        <p class="subtitle">{{horseCharacter.horse_smarts}}</p> <!-- This may need to be horse_smarts -->
+        <p class="subtitle">{{horseCharacter.horse_weight}}</p>
+        <p class="subtitle">{{horseCharacter.horse_description}}</p>
+    </div>
+</div>
+{{/each}}
