+++ views/login.handlebars
@@ -7,7 +7,7 @@
                 </header>
                 <div class="card-content">
                     <div class="content">
-                        <form action="">
+                        <form action="" class="login-form">
                             <div class="field">
                                 <label class="label" for="email">Email</label>
                                 <div class="control">
@@ -37,3 +37,5 @@
         </div>
     </div>
 </div>
+
+<script src="js/login.js"></script>
