+++ views/signup-success.handlebars
@@ -1,0 +1,17 @@
+<div class="section">
+    <div class="container">
+        <div class="columns is-centered">
+            <div class="column is-half">
+                <div class="notification is-success">
+                    <h1 class="title">Sign Up Successful!</h1>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+
+<script>
+    setTimeout(function () {
+        window.location.href = '/login';
+    }, 1500); // Redirects to login after 3 seconds
+</script>
