+++ views/signup.handlebars
@@ -7,11 +7,11 @@
                 </header>
                 <div class="card-content">
                     <div class="content">
-                        <form action="">
+                        <form action="" class="signup-form">
                             <div class="field">
                                 <label class="label" for="username">Username</label>
                                 <div class="control">
-                                    <input type="text" class="input" id="name-signup" placeholder="e.g. bobsmith"
+                                    <input type="text" class="input" id="username-signup" placeholder="e.g. bobsmith"
                                         required>
                                 </div>
                             </div>
@@ -41,3 +41,5 @@
         </div>
     </div>
 </div>
+
+<script src="js/signup.js"></script>
